Add a build script to ensure submodule dependencies are built (#43)
This adds a build script to the Visual Studio build process. The script will build each submodule dependency with the correct config and platform and put each into a directory scoped to the config and platform to avoid conflicts.

This will allow users or ci to more easily build from head.