Update the packaging of our application (#141)
The new packaging allows us to produce a single binary that can run on any debian.8 x64 machines (which is the images we will support). This mean switching over the app to target netcoreapp2.0 and produce a standalone binary for the agent. We need to upgrade to 2.0 as bug exist in 1.1 around producing standalone apps.

This change also updates our test app so it can be deployed to GAE Flex and GKE easily.