Update coverage to also cover integration tests (#153)
This is actually only partially working, due to issue #146 when things do not shut down properly. When this happens the shutdown and flush hooks in the c++ code that generate coverage don't always get called.

This switches to lcov as it is easier to combine reports (integration and unit tests), it has a much nicer UI and will make switching to codecov in the future easier.