Ensure the debugger always shuts down (#157)
The debugger does not currently shut itself down on Linux. On windows it will shut itself down when the agents pipes close. When on Linux even when the pipes are closed the debugger does not shut itself down. Trying to kill the process also does not solve the issue as the process hang indefinitely (probably waiting for the pipe).

This PR adds a field to the breakpoint that will kill the server by exiting the main program loop. The program will then always shutdown cleanly. This PR also ensures the coverage for the integration tests are accounted for.

While I don't think the current solution is the best I think this is the general model we should use (even on windows, it will give a cleaner shutdown). Possibly open up another pipe stream and write to it?

See #146, #119, #55, #27