Do not require a dll to as the only way to start the debugger (#166)
Instead of using donet and a path to a dll to start the users application to debug allow any command. This will allow users to not only specify a dll but also a stand alone executable and to also easily send other args to the application they wish to start.

Fixes #155

This change will also allow us to run integration tests in parallel on different ports.