Use a unique pipe name to communicate between agent and debugger (#173)
Currently the agent and debugger communicate via named pipe. The pipe name is not unique so if there are multiple agents running only one will read a written breakpoint. If is problematic if the two agents are not registered as the same debuggee then a breakpoint could be lost.

This also fixes the flakiness of the performance tests.