Only build submodules when needed on Linux (#178)
This is already done on Window.

This change also includes updates to travis caching (which should work better with this change) to hopefully reduce build times. We now only cache the built library directories we need. protobuf is not cached as it installs the lib on the system and is a little more complex