Fix bug in agent where it was removing non final breakpoints.
This was happening as when the hanging get to list avtice breakpoints comes back it returns a success code.  We were not checking that the wait had expried and would return an empty list of active breakpoints.

This change also adds a new sub project for long running tests.  I added tests there as they can take 10x or more longer than the regular integration tests.