Update CommandLineParser to 2.2.1 (#313)
This is a version of CommandLineParser that officially supports .NET Core Apps. It requires minor refactoring as the API has changed.

This also makes that change that we no longer throw exceptions when an error occurs parsing command line args. Instead we print the error, show help text and exit the program.

More tests are also added.